+++ datasophon-service/src/main/java/com/datasophon/api/service/impl/ClusterYarnQueueServiceImpl.java
@@ -64,12 +64,12 @@ public class ClusterYarnQueueServiceImpl extends ServiceImpl<ClusterYarnQueueMap
 
     @Override
     public Result listByPage(Integer clusterId, Integer page, Integer pageSize) {
-        Integer offset = (page - 1) * pageSize;
+        int offset = (page - 1) * pageSize;
         List<ClusterYarnQueue> list = this.list(new QueryWrapper<ClusterYarnQueue>()
                 .eq(Constants.CLUSTER_ID, clusterId)
                 .orderByDesc(Constants.CREATE_TIME)
                 .last("limit " + offset + "," + pageSize));
-        int count = this.count(new QueryWrapper<ClusterYarnQueue>()
+        long count = this.count(new QueryWrapper<ClusterYarnQueue>()
                 .eq(Constants.CLUSTER_ID, clusterId));
         for (ClusterYarnQueue clusterYarnQueue : list) {
             String minResources = clusterYarnQueue.getMinCore() + "Core," + clusterYarnQueue.getMinMem() + "GB";
